Implemented code generation for an intermediate instruction set.
The AST is flattened into a linear sequence of instructions where each
instruction can only accept previously computed values as arguments. These
are assigned to unique local variable numbers which can later be optimised
and assigned to registers or stack slots by a register allocator.

Block IDs are now allocated after the AST is created so that data types
and object types have a distinct, contiguous range of IDs. Later on this
will help the GC distinguish between tagged pointers and data.