Implemented error propagation and exception handling.
The return value of every call is checked for the presence of an error bit.
If present, the method returns immediately to the caller, returning the
error value.

Propagation can be halted with try: or catch: which clear the error bit.

This is a much simpler model than checking error values on every statement
and then jumping to the return point of the statement, which required
resolving the live set of variables and stack slots due to several
different control flow paths to the same point in the method. It is also
probably less error prone since errors are checked as soon as they are
returned and cannot be accidentally passed as arguments.