Replace allocation code with a call to OME_allocate. This is because we cannot guarantee that the registers saved to the stack contain valid pointers or tagged values since they may contain arbitrary data. The call ensures that all tagged values are stored to the stack before allocation.
This does have the downside of flushing out all registers for every
allocation even if this is not necessary most of the time. A more optimal
but more complex solution would be to keep track of the registers we need
to save and emit push instructions to save them before calling
OME_collect_nursery. This would increase the code size of every allocation
site though.